requirements.txt
streamlit
psycopg2-binary
pandas
matplotlib
plotly
python-dotenv
